#125 処理の見直し

+++ bin/cache/count.sh
@@ -40,7 +40,9 @@ update_cache() {
         fi
         return
     elif [ "$Cache_Name" = "err_mail" ] && [ "$ERR_CHK_TIME" = 0 ]; then
-        rm -f "$Cache_Name"
+        if [ -f "$Cache_File" ]; then
+            rm -f "$Cache_Name"
+        fi
         return
     fi
     # キャッシュファイルが存在するか確認
